fix(data): avoid circular dependency
