+++ test/unit/test_batch.py
@@ -4,8 +4,8 @@ import torch
 from torch.utils.data import BatchSampler, DataLoader, RandomSampler, SequentialSampler
 
 from joeynmt.batch import Batch
-from joeynmt.data import TokenBatchSampler
-from joeynmt.data_loader import load_data
+from joeynmt.data import load_data
+from joeynmt.datasets import TokenBatchSampler
 
 
 class TestBatch(TensorTestCase):
